fix# 健康通展示bug

+++ ehealth-recipe-manager/src/main/java/recipe/manager/OrderFeeManager.java
@@ -84,7 +84,7 @@ public class OrderFeeManager extends BaseManager {
             BigDecimal accountFee = getAccountFee(order.getTotalFee(), order.getMpiId(), order.getOrganId());
             if (null != accountFee) {
                 order.setThirdPayType(2);
-                order.setThirdPayFee(accountFee);
+                order.setThirdPayFee(accountFee.doubleValue());
             }
         }
     }
